Actor: Input handling
- Rotates with input
- Has rotation multiplier
- Rotation multiplier becomes zero when ragdoll activated

+++ Assets/Scripts/FFStudio/GameSettings.cs
@@ -13,7 +13,7 @@ namespace FFStudio
         [Foldout("UI Settings"), Tooltip("Duration of the fading for ui element")] public float ui_Entity_Fade_TweenDuration;
 		[Foldout("UI Settings"), Tooltip("Duration of the scaling for ui element")] public float ui_Entity_Scale_TweenDuration;
 		[Foldout("UI Settings"), Tooltip("Duration of the movement for floating ui element")] public float ui_Entity_FloatingMove_TweenDuration;
-        [Foldout("UI Settings"), Tooltip("Percentage of the screen to register a swipe")] public int swipeThreshold;
+        [Foldout("UI Settings"), Tooltip("Percentage of the screen to register a swipe")] public float swipeThreshold;
 
 
         private static GameSettings instance;
